Hack in incomplete casting + new `untyped_ptr`
This adds support for casting an expression using `expr as type`.
For now, we do not actually perform any checks on whether it is
valid to cast from a type to another.

Additionally, an `untyped_ptr` type is added to mirror `void*` in
C, to be able to use malloc/fread/etc