[breaking] Make methods denote whether `this` is passed by value
Now having `def Foo::foo(this)` means that the instance object
will be passe by _value_, instead of by reference. This is useful
for chaining, where we might not want to always store intermediate
values in separate variables.

Usually, we will want to do `def Foo::foo(&this)`, which is kind-of
inconvenient, but seems saner than having extra syntax for passing
by-value, which would seem inconsistent with the rest of the type
system.