Add `null` + make `untyped_ptr` equivalent to all pointers
This causes some behaviour that's not ideal where we can convert
back and forth between pointers of different type, but overall seems
very convenient to be able to check for null pointers of different
types without manually casting `0`.

Perhaps in the future we can make it so that only `null` can be
cast to any type, but that we cannot do so for an arbitrary
`untyped_ptr`.