Allow attaching methods to strings + refactor
This implementation is pretty hacky, but makes using strings feel
a lot nicer. The source of the issue here is that we fundamentally
expect methods to be attached only to structs, NOT primitive types.

The commit makes it so that we rely on `Type.name` being non-null
as an indicator of whether we can attach methods to something,
and then manually always set the `name` field for `&char` types to
"string".

Going forward we probably want a nicer way to be able to indicate
to the type system that a certain type (not necessarily a struct)
can have methods attached to it.