Add expression match and if statements (using GNU C extension)
This commit adds the ability to use if and match statements as
expressions. Like `jakt`, we use the `yield` keyword to return
a value from a block:

```
let x = match 0 {
  0 => {
    yield "hello"
  }
  else => {
    yield "bye"
  }
}
```

If the body of the match case or if statement is NOT a block,
but rather an expression, then it is implicitly yielded:

```
let x = match 0 {
  0 => "hello"
  else => "bye"
}
```

Note that the generated code is NOT part of the C spec, but an
extension provided by GNU C (and also clang). As such, portability
may be limited:
      https://gcc.gnu.org/onlinedocs/gcc/Statement-Exprs.html