Make expression parsing more robust
Earlier we were only able to specify "in_parens" to the expression
parser to tell it if we wanted to stop parsing at newlines or not.

Now, we can tell it what the next token we are expecting is, so
in cases where there is a known keyword / special token on both
sides of the expression we can ignore newlines completely.

We also add a dummy "TokenType::Newline", which isn't actually
emitted by the lexer, but is convenient to compare against by
looking at `.token().seen_newline`