Allow returning from yielding blocks + refactor
In many cases we wanted to throw an error and exit in the `else`
block for match statements, which was not allowed previously.
Now we just check to see if a certain branch exits the function
early and just ignore it when type checking.

Also refactor some of the code to use the new expression match.