Add support for global constants
The ideal is that constants can only be defined through compile-time
constant expressions, and so can only use literals and other constant
variables.

These have the advantage of being defined _before_ the struct
definitions in the generated C code, and so constants can be used to
size arrays / etc inside the struct types.