Re-do error handling across compiler.
This is a step towards making the compiler resistant to errors in
the code. The end goal here is to always be able to complete the
type checking phase without dying, and collect all the errors to
show the user later. This makes the code a lot uglier in places,
but seems like a necessary step towards having better IDE support
and making the compiler more resistant.

For now, there are certain instances where the parser will panic
on unexpected tokens, but this seems better than continuing parsing
garbage. This needs to be made more robust in the future so we
can remove the `.panic()` when a new error is created.