Added more inference across the compiler
This allows most integer literals across the compiler to infer
the type based on the expression / asignment they are being used
in, as opposed to having to manually specify with suffixes.

Additionally, if we are expecting an `enum` value, we can also
automatically look up an identifier in an enum, allowing you to leave
out the `enum::` prefix in most places.